リビジョンをintからlongへ
git-svn-id: https://dev.rough-diamond.jp/svn/framework/accountPattern2/core/trunk/ap2-core@838 32b2b61f-7f20-47f8-a8dc-572efa4bece0

+++ src/java/jp/rough_diamond/account/entity/base/BaseCode.java
@@ -52,7 +52,7 @@ public abstract class BaseCode  implements Serializable {
     /**
      * リビジョン
     **/ 
-    private Integer revision;
+    private Long revision;
     public final static String REVISION = "revision";
 
     /**
@@ -62,7 +62,7 @@ public abstract class BaseCode  implements Serializable {
      *    not-null="true"
      * @return リビジョン
     **/
-    public Integer getRevision() {
+    public Long getRevision() {
         return revision;
     }
 
@@ -70,7 +70,7 @@ public abstract class BaseCode  implements Serializable {
      * リビジョンを設定する
      * @param revision  リビジョン
     **/
-    public void setRevision(Integer revision) {
+    public void setRevision(Long revision) {
         this.revision = revision;
     }
     /**
